OK. That's figured out.
The JSON module error has to do with the fact that I have split
the module in multiple parts and use import when running.
This just to prevent duplicating code.
This works very well when testing directory with Python and
with Ansible version 2.7 and 2.8. But with 2.9 things go bad.

I created a new `mm_dhcp.py` with everything in it and that works.
I need to figure out a way to develop using the import so when
things change I do not need to fix this in a lot of places.

Need to think of that.