QF-01: Force basic authentication
When mmweb is configured for SSO, all API authentication calls are
failing with error "AbstractDigestAuthHandler does not support the
following scheme: 'Negotiate'".

We should automatically failback to basic authentication, but since that
is not happening automatically we need to force basic authentication in
the 'open_url' function call.

This fix came from Ton Kersten.